structure files accordingly
